requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터 처리 및 검증
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 템플릿 및 파일 처리
jinja2==3.1.2
aiofiles==23.2.1

# LLM 및 AI
openai>=1.0.0

# 보안 및 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 모니터링 및 로깅
prometheus-client==0.19.0

# 데이터베이스 (선택 사항)
psycopg2-binary==2.9.9  # PostgreSQL 사용 시